% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{alert}
\alias{alert}
\title{Create a Tabler-styled dismissible alert}
\usage{
alert(title, message, type = "warning")
}
\arguments{
\item{title}{Alert title}

\item{message}{Alert message}

\item{type}{Alert type ("warning", "info", "success", "danger")}
}
\value{
A shiny tag list
}
\description{
Create a Tabler-styled dismissible alert
}
