% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_edit_url}
\alias{generate_edit_url}
\title{Generate edit URL}
\usage{
generate_edit_url(submission_id)
}
\arguments{
\item{submission_id}{Character. ID of the submission to edit}
}
\value{
Character. Edit URL or NULL if generation fails
}
\description{
Generates an edit URL for a KoboToolbox submission
}
