% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler_header.R
\name{header}
\alias{header}
\title{Add a header}
\usage{
header(..., logo = NULL)
}
\arguments{
\item{...}{html tags to be included on the right hand side of the header}

\item{logo}{html for the logo}
}
\description{
Add a header
}
