% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hex_map.R
\name{heatmap_map}
\alias{heatmap_map}
\title{Create a heatmap using deckgl}
\usage{
heatmap_map(data)
}
\arguments{
\item{data}{A dataframe containing lat/lon coordinates}
}
\value{
A deckgl object
}
\description{
Create a heatmap using deckgl
}
