% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hex_map.R
\name{hexagon_map}
\alias{hexagon_map}
\title{Create a hexagon heatmap using deckgl}
\usage{
hexagon_map(data, radius = 600)
}
\arguments{
\item{data}{A dataframe containing lat/lon coordinates}

\item{radius}{Numeric. Radius of hexagons in meters}
}
\value{
A deckgl object
}
\description{
Create a hexagon heatmap using deckgl
}
