% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_plots.R
\name{homecard_ts_plot_memo}
\alias{homecard_ts_plot_memo}
\title{Memoised version of homecard_ts_plot}
\usage{
homecard_ts_plot_memo(data, x_col = "date_month", y_col, type = "area")
}
\arguments{
\item{data}{A dataframe containing the time series data}

\item{x_col}{The name of the column to be used for the x-axis (default: "date_month")}

\item{y_col}{The name of the column to be used for the y-axis}

\item{type}{The type of plot (default: "area")}
}
\description{
Memoised version of homecard_ts_plot
}
