% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_data}
\alias{length_data}
\title{Fish Length Data}
\format{
A tibble with columns:
  \describe{
    \item{catch_taxon}{Taxonomic classification}
    \item{q25}{25th percentile of length distribution}
    \item{q75}{75th percentile of length distribution}
    \item{min}{Minimum recorded length}
    \item{max}{Maximum recorded length}
  }
}
\usage{
length_data
}
\description{
Length statistics for the top 11 taxa by catch weight
}
\details{
Length measurements are weighted by count frequency. Taxa names are
  simplified for some groups (e.g., "Tuna/Bonito/Other Mackerel" to "Tuna/Bonito")
}
\keyword{datasets}
