% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_data}
\alias{map_data}
\title{Fishing Location Data}
\format{
A data frame with columns:
  \describe{
    \item{lat}{Latitude coordinate}
    \item{lon}{Longitude coordinate}
  }
}
\usage{
map_data
}
\description{
Geographical coordinates of valid fishing activities, cleaned to remove NA values
}
\details{
Data has been cleaned to remove any NA values in coordinates
}
\keyword{datasets}
