% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mozamap_with_counts}
\alias{mozamap_with_counts}
\title{District Boundaries with Survey Counts}
\format{
A simple feature collection with columns:
  \describe{
    \item{ADM2_PT}{District name in Portuguese}
    \item{ADM2_PCODE}{District administrative code}
    \item{ADM1_PT}{Province name in Portuguese (Cabo Delgado)}
    \item{ADM1_PCODE}{Province administrative code}
    \item{ADM0_EN}{Country name in English (Mozambique)}
    \item{ADM0_PT}{Country name in Portuguese (Moçambique)}
    \item{n_observations}{Number of survey points falling within each district}
    \item{geometry}{MULTIPOLYGON geometry defining district boundaries}
  }
}
\usage{
mozamap_with_counts
}
\description{
Spatial data for Palma and Mocimboa Da Praia districts including the count of survey points within each district
}
\details{
Districts are filtered to include only Palma and Mocimboa Da Praia.
  Survey points are counted using spatial intersection with district polygons.
}
\keyword{datasets}
