% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler_cards.R
\name{small_card}
\alias{small_card}
\title{Create a small card with an embedded plot that fills the entire card and ensures visible tooltip}
\usage{
small_card(title = "Card title", plot)
}
\arguments{
\item{title}{The title of the card}

\item{plot}{The apexchart object to embed}
}
\value{
A shiny tag
}
\description{
Create a small card with an embedded plot that fills the entire card and ensures visible tooltip
}
