% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab_data}
\alias{tab_data}
\title{Aggregated Landing Site Data}
\format{
A tibble with columns:
  \describe{
    \item{district}{District name}
    \item{landing_site}{Name of the landing site}
    \item{trip_duration_hrs}{Mean trip duration in hours}
    \item{cpue_kg_fisher_hr}{Mean catch per unit effort (kg/fisher/hour)}
    \item{price_per_kg_mzn}{Mean price per kilogram in Mozambican Metical}
    \item{mean_catch_kg}{Mean catch in kilograms}
    \item{mean_catch_price_mzn}{Mean catch price in Mozambican Metical}
  }
}
\usage{
tab_data
}
\description{
Summary statistics for fishing activities by district and landing site
}
\keyword{datasets}
