% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_revenue.R
\name{tab_revenue_content}
\alias{tab_revenue_content}
\title{Generate content for the revenue tab of the Malawi small scale fisheries dashboard}
\usage{
tab_revenue_content()
}
\value{
A tagList containing the structured content for the revenue tab
}
\description{
This function creates the layout and content for the revenue tab of the dashboard.
It includes a time series plot and a revenue distribution plot.
}
