% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_df}
\alias{taxa_df}
\title{Taxa Catch Data}
\format{
A tibble with columns:
  \describe{
    \item{submission_id}{Unique identifier for each submission}
    \item{catch_taxon}{Taxonomic classification of the catch}
    \item{catch_gr}{Catch weight in grams}
  }
}
\usage{
taxa_df
}
\description{
Detailed catch data by taxonomic group
}
\keyword{datasets}
