% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_summary}
\alias{taxa_summary}
\title{Taxa Summary by Landing Site}
\format{
A tibble with columns:
  \describe{
    \item{landing_site}{Name of the landing site}
    \item{catch_taxon}{Taxonomic classification (top 6 taxa and "Other")}
    \item{catch_kg}{Total catch in kilograms}
    \item{catch_percent}{Percentage of total catch}
  }
}
\usage{
taxa_summary
}
\description{
Summary of catch composition by taxa at each landing site
}
\details{
Top 6 taxa are identified based on total catch weight, with remaining taxa
  grouped as "Other"
}
\keyword{datasets}
